+++ CHANGELOG.md
@@ -4,7 +4,14 @@ All notable changes to this project will be documented in this file.
 
 Most patterns are highly stable, no changes will be made to existing methods, only extended, but I will adhere to adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) just in case. You can safely upgrade, but as always, RTFM (read changelog for major releases).
 
-# [2.1.3] - 2025-06-21
+
+# [2.1.4] - 2025-06-21
+
+### Changed:
+
+- RealtimeChannel now use publish, instead of emit. Emit is depricated.
+
+# [2.1.3] - 2025-06-21 {# [2.1.3] - 2025-06-21  ### Added:  - EventChannel - EventBrokerServer}
 
 ### Added: